--- csharp-advanced-topics/ActionAndFuncDelegatesInCsharp/ActionAndFuncDelegatesInCsharp.sln
@@ -1,13 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.6.33815.320
-MinimumVisualStudioVersion = 10.0.40219.1
-Global
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {75B6C8A3-B0A0-4B8C-B2F2-860CA58375D1}
-	EndGlobalSection
-EndGlobal
