+++ csharp-advanced-topics/ActionDelegatesInCSharp/ActionDelegatesInCSharp.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33815.320
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionDelegatesInCSharp", "ActionDelegatesInCSharp\ActionDelegatesInCSharp.csproj", "{2CD382FE-347E-41AD-BD76-0A8C251F17CD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ActionDelegatesInCsharpUnitTest", "ActionDelegatesInCsharpUnitTest\ActionDelegatesInCsharpUnitTest.csproj", "{202A5721-AE73-4930-A462-1A234AEF081D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2CD382FE-347E-41AD-BD76-0A8C251F17CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2CD382FE-347E-41AD-BD76-0A8C251F17CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2CD382FE-347E-41AD-BD76-0A8C251F17CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2CD382FE-347E-41AD-BD76-0A8C251F17CD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{202A5721-AE73-4930-A462-1A234AEF081D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{202A5721-AE73-4930-A462-1A234AEF081D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{202A5721-AE73-4930-A462-1A234AEF081D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{202A5721-AE73-4930-A462-1A234AEF081D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E210304B-D6F7-40F6-B9A2-E20B7633BF2A}
+	EndGlobalSection
+EndGlobal
