+++ csharp-advanced-topics/ActionDelegatesInCSharp/ActionDelegatesInCSharp/Classes/Sounds.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿
 
-namespace AnimalSound.Classes
+namespace ActionDelegatesInCSharp.Classes
 {
     public class Sounds
     {
@@ -16,4 +12,5 @@ namespace AnimalSound.Classes
 
 
     }
+
 }
