+++ csharp-advanced-topics/ActionDelegatesInCSharp/ActionDelegatesInCsharpUnitTest/UnitTest1.cs
@@ -1,0 +1,11 @@
+namespace ActionDelegatesInCsharpUnitTest
+{
+    [TestClass]
+    public class UnitTest1
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
