+++ csharp-advanced-topics/ActionDelegatesInCSharp/ActionDelegatesInCsharpUnitTest/Usings.cs
@@ -1,0 +1,1 @@
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
